+++ lib/datasource/remote/polkadot_api/polkadot_repository.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'dart:math';
 
 import 'package:hashed/datasource/local/account_service.dart';
-import 'package:hashed/datasource/local/flutter_js/polkawallet_init.dart';
+import 'package:hashed/datasource/local/flutter_js/substrate_service.dart';
 import 'package:hashed/datasource/remote/model/guardians_config_model.dart';
 import 'package:hashed/datasource/remote/model/token_model.dart';
 import 'package:hashed/datasource/remote/polkadot_api/recovery_repository.dart';
